requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
PyJWT==2.8.0
PyJWT==2.10.1
python-dotenv==1.0.1
Requests==2.32.3
